+++ Source/Scene/PostProcessStageLibrary.js
@@ -1,6 +1,8 @@
 define([
         '../Core/buildModuleUrl',
+        '../Core/createGuid',
         '../Core/Color',
+        '../Core/defined',
         '../Core/defineProperties',
         '../Core/destroyObject',
         '../Core/Ellipsoid',
@@ -16,7 +18,6 @@ define([
         '../Shaders/PostProcessStages/FXAA',
         '../Shaders/PostProcessStages/GaussianBlur1D',
         '../Shaders/PostProcessStages/LensFlare',
-        '../Shaders/PostProcessStages/LinearDepth',
         '../Shaders/PostProcessStages/NightVision',
         '../Shaders/PostProcessStages/Silhouette',
         '../ThirdParty/Shaders/FXAA3_11',
@@ -25,7 +26,9 @@ define([
         './PostProcessStageSampleMode'
     ], function(
         buildModuleUrl,
+        createGuid,
         Color,
+        defined,
         defineProperties,
         destroyObject,
         Ellipsoid,
@@ -41,7 +44,6 @@ define([
         FXAA,
         GaussianBlur1D,
         LensFlare,
-        LinearDepth,
         NightVision,
         Silhouette,
         FXAA3_11,
@@ -215,9 +217,9 @@ define([
     };
 
     /**
-     * Creates a post-process stage that applies a silhouette effect.
+     * Creates a post-process stage that detects edges.
      * <p>
-     * A silhouette effect highlights the edges of an object.
+     * Writes the color to the output texture with alpha set to 1.0 when it is on an edge.
      * </p>
      * <p>
      * This stage has the following uniforms: <code>color</code> and <code>length</code>
@@ -226,57 +228,112 @@ define([
      * <li><code>color</code> is the color of the highlighted edge. The default is {@link Color#BLACK}.</li>
      * <li><code>length</code> is the length of the edges in pixels. The default is <code>0.5</code>.</li>
      * </ul>
-     * @return {PostProcessStageComposite} A post-process stage that applies a silhouette effect.
+     * @return {PostProcessStageComposite} A post-process stage that applies an edge detection effect.
+     *
+     * @example
+     * // multiple silhouette effects
+     * var yellowEdge = Cesium.PostProcessLibrary.createEdgeDetectionStage();
+     * yellowEdge.uniforms.color = Cesium.Color.YELLOW;
+     * yellowEdge.selectedFeatures = [feature0];
+     *
+     * var greenEdge = Cesium.PostProcessLibrary.createEdgeDetectionStage();
+     * greenEdge.uniforms.color = Cesium.Color.LIME;
+     * greenEdge.selectedFeatures = [feature1];
+     *
+     * // draw edges around feature0 and feature1
+     * postProcessStages.add(Cesium.PostProcessLibrary.createSilhouetteEffect([yellowEdge, greenEdge]);
      */
-    PostProcessStageLibrary.createSilhouetteStage = function() {
-        var silhouetteDepth = new PostProcessStage({
-            name : 'czm_silhouette_depth',
-            fragmentShader : LinearDepth
-        });
-        var edgeDetection = new PostProcessStage({
-            name : 'czm_silhouette_edge_detection',
+    PostProcessStageLibrary.createEdgeDetectionStage = function() {
+        // unique name generated on call so more than one effect can be added
+        var name = createGuid();
+        return new PostProcessStage({
+            name : 'czm_edge_detection_' + name,
             fragmentShader : EdgeDetection,
             uniforms : {
                 length : 0.25,
                 color : Color.clone(Color.BLACK)
             }
         });
-        var silhouetteGenerateProcess = new PostProcessStageComposite({
-            name : 'czm_silhouette_generate',
-            stages : [silhouetteDepth, edgeDetection]
+    };
+
+    function getSilhouetteEdgeDetection(edgeDetectionStages) {
+        if (!defined(edgeDetectionStages)) {
+            return PostProcessStageLibrary.createEdgeDetectionStage();
+        }
+
+        var edgeDetection = new PostProcessStageComposite({
+            name : 'czm_edge_detection_multiple',
+            stages : edgeDetectionStages,
+            inputPreviousStageTexture : false
         });
+
+        var compositeUniforms = {};
+        var fsDecl = '';
+        var fsLoop = '';
+        for (var i = 0; i < edgeDetectionStages.length; ++i) {
+            fsDecl += 'uniform sampler2D edgeTexture' + i + '; \n';
+            fsLoop +=
+                '        vec4 edge' + i + ' = texture2D(edgeTexture' + i + ', v_textureCoordinates); \n' +
+                '        if (edge' + i + '.a > 0.0) \n' +
+                '        { \n' +
+                '            color = edge' + i + '; \n' +
+                '            break; \n' +
+                '        } \n';
+            compositeUniforms['edgeTexture' + i] = edgeDetectionStages[i].name;
+        }
+
+        var fs =
+            fsDecl +
+            'varying vec2 v_textureCoordinates; \n' +
+            'void main() { \n' +
+            '    vec4 color = vec4(0.0); \n' +
+            '    for (int i = 0; i < ' + edgeDetectionStages.length + '; i++) \n' +
+            '    { \n' +
+            fsLoop +
+            '    } \n' +
+            '    gl_FragColor = color; \n' +
+            '} \n';
+
+        var edgeComposite = new PostProcessStage({
+            name : 'czm_edge_detection_combine',
+            fragmentShader : fs,
+            uniforms : compositeUniforms
+        });
+        return new PostProcessStageComposite({
+            name : 'czm_edge_detection_composite',
+            stages : [edgeDetection, edgeComposite]
+        });
+    }
+
+    /**
+     * Creates a post-process stage that applies a silhouette effect.
+     * <p>
+     * A silhouette effect composites the color from the edge detection pass with input color texture.
+     * </p>
+     * <p>
+     * This stage has the following uniforms when <code>edgeDetectionStages</code> is <code>undefined</code>: <code>color</code> and <code>length</code>
+     * </p>
+     * <p>
+     * <code>color</code> is the color of the highlighted edge. The default is {@link Color#BLACK}.
+     * <code>length</code> is the length of the edges in pixels. The default is <code>0.5</code>.
+     * </p>
+     * @return {PostProcessStageComposite} A post-process stage that applies a silhouette effect.
+     */
+    PostProcessStageLibrary.createSilhouetteStage = function(edgeDetectionStages) {
+        var edgeDetection = getSilhouetteEdgeDetection(edgeDetectionStages);
         var silhouetteProcess = new PostProcessStage({
             name : 'czm_silhouette_color_edges',
             fragmentShader : Silhouette,
             uniforms : {
-                silhouetteTexture : silhouetteGenerateProcess.name
+                silhouetteTexture : edgeDetection.name
             }
         });
 
-        var uniforms = {};
-        defineProperties(uniforms, {
-            length : {
-                get : function() {
-                    return edgeDetection.uniforms.length;
-                },
-                set : function(value) {
-                    edgeDetection.uniforms.length = value;
-                }
-            },
-            color : {
-                get : function() {
-                    return edgeDetection.uniforms.color;
-                },
-                set : function(value) {
-                    edgeDetection.uniforms.color = value;
-                }
-            }
-        });
         return new PostProcessStageComposite({
             name : 'czm_silhouette',
-            stages : [silhouetteGenerateProcess, silhouetteProcess],
+            stages : [edgeDetection, silhouetteProcess],
             inputPreviousStageTexture : false,
-            uniforms : uniforms
+            uniforms : edgeDetection.uniforms
         });
     };